backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.3
python-docx==1.1.0
pdf2image==1.17.0
pillow>=9.0.0,<11.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
pypandoc==1.12
pdf2docx==0.5.6
PyMuPDF==1.23.22 